+++ NetDataTransfer-android/app/src/main/AndroidManifest.xml
@@ -86,10 +86,10 @@
             </intent-filter>
         </activity>
 
-        <service android:name="net.service.BroadcastMonitorService" />
+        <service android:name="net.service.LoginMonitorService" />
         <service android:name="net.service.UdpDataMonitorService" />
         <service android:name="net.service.FileMonitorService" />
-        <service android:name="net.service.ScreenMonitorService" />
+        <service android:name="net.service.BroadcastMonitorService" />
     </application>
 
 </manifest>
