+++ NetDataTransfer-android/app/src/main/java/net/ui/activity/BaseActivity.java
@@ -1,19 +1,28 @@
 package net.ui.activity;
 
 import android.app.Activity;
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+import android.os.Message;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
+import android.widget.Toast;
 
-import net.service.BroadcastMonitorService;
+import net.app.NetConfApplication;
+import net.service.LoginMonitorService;
 import net.service.FileMonitorService;
-import net.service.ScreenMonitorService;
+import net.service.BroadcastMonitorService;
 import net.service.UdpDataMonitorService;
 
 import java.lang.reflect.Field;
 
-public class BaseActivity extends Activity {
+public class BaseActivity extends Activity implements NetConfApplication.WifiListener {
+    NetConfApplication app;
+
     /**
      * 简化findViewById操作
      */
@@ -22,10 +31,68 @@ public class BaseActivity extends Activity {
     }
 
     @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        app = (NetConfApplication) getApplication();
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        app.listeners.add(this);
+        if (app.wifi != 1 && NetConfApplication.isUIReady) {
+            warnWifiState();
+        }
+    }
+
+    @Override
+    protected void onPause() {
+        app.listeners.remove(this);
+        super.onPause();
+    }
+
+    @Override
     protected void onDestroy() {
         fixInputMethodManagerLeak(this);
         super.onDestroy();
     }
+
+    @Override
+    public void notifyWifiInfo() {
+        warnWifiState();
+    }
+
+    public void warnWifiState() {
+        if (app.wifi == 1) {
+            Toast.makeText(this, "网络已连接，请继续使用", Toast.LENGTH_SHORT).show();
+        } else {
+            new AlertDialog.Builder(this).setTitle("错误")
+                    .setMessage("wifi已断开")
+                    .setPositiveButton("退出",
+                            new DialogInterface.OnClickListener() {
+                                @Override
+                                public void onClick(DialogInterface dialog, int which) {
+                                    setResult(RESULT_OK);// 确定按钮事件
+                                    System.exit(0);
+                                }
+                            })
+                    .setNegativeButton("重试",
+                            new DialogInterface.OnClickListener() {
+                                @Override
+                                public void onClick(DialogInterface dialog, int which) {
+                                    handler.sendEmptyMessageDelayed(app.wifi, 1000);
+                                }
+                            }).setCancelable(false).show();
+        }
+    }
+
+    Handler handler = new Handler() {
+        @Override
+        public void handleMessage(Message msg) {
+            app.check(false);
+            warnWifiState();
+        }
+    };
 
     public void fixInputMethodManagerLeak(Context destContext) {
         if (destContext == null) {
@@ -64,9 +131,9 @@ public class BaseActivity extends Activity {
     }
 
     protected void listen() {
-        this.startService(new Intent(this, BroadcastMonitorService.class));
+        this.startService(new Intent(this, LoginMonitorService.class));
         this.startService(new Intent(this, UdpDataMonitorService.class));
         this.startService(new Intent(this, FileMonitorService.class));
-        this.startService(new Intent(this, ScreenMonitorService.class));
+        this.startService(new Intent(this, BroadcastMonitorService.class));
     }
 }
